requirements.txt
africastalking==1.2.4
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
distlib==0.3.5
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.13.1
django_heroku
django-countries==7.3.2
django-phonenumber-field==6.3.0
drf-yasg==1.21.3
filelock==3.8.0
gunicorn==20.1.0
identify==2.5.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.53
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
virtualenv==20.16.3
wcwidth==0.2.5
wrapt==1.14.1
